+++ src/assets/data/trending-courses.json
@@ -2,7 +2,7 @@
     {
         "id": 3,
         "name": "JavaScript Basics",
-        "cover": "https://unsplash.com/photos/TEemXOpR3cQ",
+        "cover": "https://testasmpublic-14e65.kxcdn.com/1658217463_3666e5dfb262d265d43fb0f905644a_file_example_PNG_500kB.png",
         "rating": 4,
         "reviews": [
         ]
@@ -10,7 +10,7 @@
     {
         "id": 7,
         "name": "Angular Basics",
-        "cover": "https://www.pexels.com/photo/road-landscape-sunset-field-7953208/",
+        "cover": "https://testasmpublic-14e65.kxcdn.com/1658217463_c2238b769f2f3ecd09461692f9846f_file_example_JPG_100kB.jpeg",
         "rating": 5,
         "reviews":[
         ]
@@ -18,7 +18,7 @@
     {
         "id": 11,
         "name": "Angular Advance",
-        "cover": "https://www.pexels.com/photo/black-screen-with-code-4164418/",
+        "cover": "https://testasmpublic-14e65.kxcdn.com/1652993116_f90e1abe063f8a1e83e2b021f456e3_logo.png",
         "rating": 4.5,
         "reviews":[
         ]
